+++ src/Scaffolder/src/Command/ControllerCommand.php
@@ -34,6 +34,12 @@ class ControllerCommand extends AbstractCommand
             InputOption::VALUE_OPTIONAL,
             'Optional comment to add as class header',
         ],
+        [
+            'namespace',
+            null,
+            InputOption::VALUE_OPTIONAL,
+            'Optional, specify a custom namespace',
+        ],
     ];
 
     /**